refactor: Refactor the front-end structure
