+++ web/package.json
@@ -14,6 +14,7 @@
     "element-plus": "^2.2.17",
     "ts-md5": "^1.3.1",
     "vue": "^3.2.37",
+    "vue-router": "^4.1.6",
     "vue3-slide-verify": "^1.1.4"
   },
   "devDependencies": {
